--- src/main/java/com/yoonje/authenticationexample/AdminInitializer.java
@@ -1,4 +1,0 @@
-package com.yoonje.authenticationexample;
-
-public class AdminInitializer {
-}
