--- src/main/java/com/yoonje/authenticationexample/dto/member/response/MemberInfoResponse.java
@@ -1,4 +1,0 @@
-package com.yoonje.authenticationexample.dto.member.response;
-
-public record MemberInfoResponse() {
-}
