--- src/main/java/com/yoonje/authenticationexample/dto/member/response/MemberUpdateResponse.java
@@ -1,4 +1,0 @@
-package com.yoonje.authenticationexample.dto.member.response;
-
-public record MemberUpdateResponse() {
-}
