--- src/main/java/com/yoonje/authenticationexample/dto/sign_in/response/SignInResponse.java
@@ -1,13 +1,0 @@
-package com.yoonje.authenticationexample.dto.sign_in.response;
-
-import com.yoonje.authenticationexample.common.MemberType;
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public record SignInResponse(
-        @Schema(description = "회원 이름", example = "윤제")
-        String name,
-        @Schema(description = "회원 유형", example = "USER")
-        MemberType type,
-        String token
-) {
-}
