--- src/main/java/com/yoonje/authenticationexample/handler/ExceptionResponseHandler.java
@@ -1,4 +1,0 @@
-package com.yoonje.authenticationexample.handler;
-
-public class ExceptionResponseHandler {
-}
