+++ src/main/java/com/yoonje/authenticationexample/member/dto/ApiResponse.java
@@ -1,0 +1,17 @@
+package com.yoonje.authenticationexample.member.dto;
+
+import com.yoonje.authenticationexample.common.ApiStatus;
+
+public record ApiResponse(
+        ApiStatus status,
+        String message,
+        Object data
+) {
+    public static ApiResponse success(Object data) {
+        return new ApiResponse(ApiStatus.SUCCESS, null, data);
+    }
+
+    public static ApiResponse error(String message) {
+        return new ApiResponse(ApiStatus.ERROR, message, null);
+    }
+}
