+++ src/main/java/com/yoonje/authenticationexample/member/dto/member/response/MemberInfoResponse.java
@@ -1,0 +1,4 @@
+package com.yoonje.authenticationexample.member.dto.member.response;
+
+public record MemberInfoResponse() {
+}
