+++ src/main/java/com/yoonje/authenticationexample/member/exception/MemberNotFoundException.java
@@ -1,0 +1,14 @@
+package com.yoonje.authenticationexample.member.exception;
+
+public class MemberNotFoundException extends RuntimeException{
+    public MemberNotFoundException() {
+    }
+
+    public MemberNotFoundException(String message) {
+        super(message);
+    }
+
+    public MemberNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
