+++ src/main/java/com/yoonje/authenticationexample/member/service/SignService.java
@@ -1,0 +1,10 @@
+package com.yoonje.authenticationexample.member.service;
+
+import com.yoonje.authenticationexample.member.dto.sign_up.request.SignUpRequest;
+
+public interface SignService {
+    void signUp(SignUpRequest request);
+    void signIn(SignUpRequest request);
+
+
+}
