--- src/main/java/com/yoonje/authenticationexample/repository/MemberRepository.java
@@ -1,16 +1,0 @@
-package com.yoonje.authenticationexample.repository;
-
-import com.yoonje.authenticationexample.common.MemberType;
-import com.yoonje.authenticationexample.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, UUID> {
-    Optional<Member> findByEmail(String email);
-    List<Member> findAllByType(MemberType type);
-}
