+++ src/main/java/com/yoonje/authenticationexample/security/JwtAuthenticationFilter.java
@@ -1,4 +1,60 @@
 package com.yoonje.authenticationexample.security;
 
-public class JwtAuthenticationFilter {
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.annotation.Order;
+import org.springframework.http.HttpHeaders;
+import org.springframework.security.authentication.AbstractAuthenticationToken;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.web.authentication.WebAuthenticationDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Optional;
+
+@Order(0)
+@RequiredArgsConstructor
+@Component
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+    private final TokenProvider tokenProvider;
+
+    // 밑의 두 메소드를 이용해 스프링 시큐리티에서 사용할 UsernamePasswordAuthenticationToken 객체를 생성한다.
+    // 이후 스프링 시큐리티 컨텍스트의 인증 정보를 새로 생성한 인증 토큰으로 설정하고 다음 필터로 넘어간다
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        String token = parseBearerToken(request);
+        User user = parseUserSpecification(token);
+        AbstractAuthenticationToken authenticated = UsernamePasswordAuthenticationToken.authenticated(user, token, user.getAuthorities());
+        authenticated.setDetails(new WebAuthenticationDetails(request));
+        SecurityContextHolder.getContext().setAuthentication(authenticated);
+
+        filterChain.doFilter(request, response);
+    }
+
+    // HTTP 요청의 헤더에서 Authrozation 값을 찾아서 Bearer로 시작하는지 확인. 그 후 접두어를 제외한 토큰값으로 파싱
+    private String parseBearerToken(HttpServletRequest request) {
+        return Optional.ofNullable(request.getHeader(HttpHeaders.AUTHORIZATION))
+                .filter(token -> token.substring(0, 7).equalsIgnoreCase("Bearer "))
+                .map(token -> token.substring(7))
+                .orElse(null);
+    }
+
+    // 토큰 값의 회원ID와 타입을 토대로 스프링 시큐리티에서 사용할 User객체 반환
+    private User parseUserSpecification(String token) {
+        String[] split = Optional.ofNullable(token)
+                .filter(subject -> subject.length() >= 10)
+                .map(tokenProvider::validateTokenAndGetSubject)
+                .orElse("anonymous:anonymous")
+                .split(":");
+
+        return new User(split[0], "", List.of(new SimpleGrantedAuthority(split[1])));
+    }
 }
